requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
pandas
openpyxl
python-dotenv
requests
Werkzeug
